hackathon/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
python-dotenv==1.0.1
sqlmodel==0.0.14
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.21.0
lxml==5.2.2
aiohttp==3.9.5
PyJWT==2.8.0
email-validator==2.1.1
